+++ sqlparser/sqlparser.h
@@ -429,10 +429,10 @@ public:
 	Token* GetNextCharToken(Token *prev, const char ch, const wchar_t wch);
 	Token* GetNextCharOrLastToken(const char ch, const wchar_t wch);
 	Token* GetNextPlusMinusAsOperatorToken(const char ch, const wchar_t wch);
-	Token* GetNextSpecialCharToken(Token *prev, const char *str, const wchar_t *wstr, int len);
-	Token* GetNextSpecialCharToken(const char *str, const wchar_t *wstr, int len);
-	Token* GetNextWordToken(const char *str, const wchar_t *wstr, int len);
-	Token* GetNextWordToken(Token *prev, const char *str, const wchar_t *wstr, int len);
+	Token* GetNextSpecialCharToken(Token *prev, const char *str, const wchar_t *wstr, size_t len);
+	Token* GetNextSpecialCharToken(const char *str, const wchar_t *wstr, size_t len);
+	Token* GetNextWordToken(const char *str, const wchar_t *wstr, size_t len);
+	Token* GetNextWordToken(Token *prev, const char *str, const wchar_t *wstr, size_t len);
 	Token* GetNextUntilNewlineToken();
 	Token* GetNextNumberToken();
 	Token* GetNextNumberToken(Token *prev);
@@ -446,8 +446,8 @@ public:
 
 	Token* GetNext() { return GetNextToken(); }
 	Token* GetNext(Token *prev) { return GetNextToken(prev); }
-	Token* GetNext(const char *str, const wchar_t *wstr, int len) { return GetNextWordToken(str, wstr, len); }
-	Token* GetNext(Token *prev, const char *str, const wchar_t *wstr, int len) { return GetNextWordToken(prev, str, wstr, len); }
+	Token* GetNext(const char *str, const wchar_t *wstr, size_t len) { return GetNextWordToken(str, wstr, len); }
+	Token* GetNext(Token *prev, const char *str, const wchar_t *wstr, size_t len) { return GetNextWordToken(prev, str, wstr, len); }
 	Token* GetNext(const char ch, const wchar_t wch) { return GetNextCharToken(ch, wch); }
 	Token* GetNext(Token *prev, const char ch, const wchar_t wch) { return GetNextCharToken(prev, ch, wch); }
 
@@ -459,12 +459,12 @@ public:
 	void PushBack(Token *token);
 
 	// Append the token with the specified value
-	Token *Append(Token *token, const char *str, const wchar_t *wstr, int len, Token *format = NULL);
+	Token *Append(Token *token, const char *str, const wchar_t *wstr, size_t len, Token *format = NULL);
 	void Append(Token *token, int value);
 	void Append(Token *token, TokenStr *str, Token *format = NULL);
-	Token* AppendNoFormat(Token *token, const char *str, const wchar_t *wstr, int len);
+	Token* AppendNoFormat(Token *token, const char *str, const wchar_t *wstr, size_t len);
 	void AppendNoFormat(Token *token, TokenStr *str);
-	void AppendFirstNoFormat(Token *token, const char *str, const wchar_t *wstr, int len);
+	void AppendFirstNoFormat(Token *token, const char *str, const wchar_t *wstr, size_t len);
 	Token* AppendCopy(Token *token, Token *append);
 	Token* AppendCopy(Token *token, Token *first, Token *last, bool append_removed = true);
 	void AppendSpaceCopy(Token *token, Token *append);
@@ -472,15 +472,15 @@ public:
 	void Append(Token *token, Token *append);
 	
 	// Prepend the token with the specified value
-	Token* Prepend(Token *token, const char *str, const wchar_t *wstr, int len, Token *format = NULL);
-	Token* PrependNoFormat(Token *token, const char *str, const wchar_t *wstr, int len);
+	Token* Prepend(Token *token, const char *str, const wchar_t *wstr, size_t len, Token *format = NULL);
+	Token* PrependNoFormat(Token *token, const char *str, const wchar_t *wstr, size_t len);
 	Token* PrependNoFormat(Token *token, TokenStr *str);
 	Token* PrependCopy(Token *token, Token *prepend);
 	Token* PrependCopy(Token *token, Token *first, Token *last, bool prepend_removed = true);
 	void PrependSpaceCopy(Token *token, Token *first, Token *last, bool prepend_removed = true);
 	void Prepend(Token *token, Token *prepend);
 
-	void ChangeWithSpacesAround(Token *token, const char *new_str, const wchar_t *new_wstr, int len, Token *format = NULL);
+	void ChangeWithSpacesAround(Token *token, const char *new_str, const wchar_t *new_wstr, size_t len, Token *format = NULL);
 	
 	// Check for the source and target type
 	bool Source(short s1, short s2 = -1, short s3 = -1, short s4 = -1, short s5 = -1, short s6 = -1); 
@@ -1218,7 +1218,7 @@ public:
 	bool ParseInformixReturns(Token *returns);
 	void InformixReturnToAssignment(ListWM &return_list);
 	void InformixConvertReturning(Token *create, Token *procedure);
-	void InformixGetReturningName(Token *retvalue, TokenStr &name, const char *prefix, const wchar_t *wprefix, int plen, int num, bool always_prefix);
+	void InformixGetReturningName(Token *retvalue, TokenStr &name, const char *prefix, const wchar_t *wprefix, size_t plen, int num, bool always_prefix);
 	bool ParseInformixIntoTable(Token *select);
 	bool InformixPatternAssignLastSerial(Token *let);
 	bool ParseInformixSetOptions(Token *set);
@@ -1253,8 +1253,8 @@ public:
 	void ConvertParameterIdentifier(Token *ref, Token *decl);
 	void ConvertVariableIdentifier(Token *token);
 	void ConvertVariableIdentifier(Token *ref, Token *decl);
-	void ConvertSchemaName(Token *token, TokenStr &ident, int *len);
-	void ConvertObjectName(Token *token, TokenStr &ident, int *len);
+	void ConvertSchemaName(Token *token, TokenStr &ident, size_t *len);
+	void ConvertObjectName(Token *token, TokenStr &ident, size_t *len);
 	bool ConvertCursorParameter(Token *token);
 	bool ConvertRecordVariable(Token *token);
 	bool ConvertImplicitForRecordVariable(Token *token);
@@ -1263,7 +1263,7 @@ public:
 	bool ConvertLocalTableColumn(Token *token);
 	bool ConvertSessionTemporaryTable(Token *token);
 	int GetIdentPartsCount(Token *token);
-	void GetNextIdentItem(Token *token, TokenStr &ident, int *len);
+	void GetNextIdentItem(Token *token, TokenStr &ident, size_t *len);
 	void SplitIdentifierByLastPart(Token *token, TokenStr &lead, TokenStr &trail, int parts = -1);
 	bool ConvertTsqlVariable(Token *token);
 	bool ConvertToTsqlVariable(Token *token);
@@ -1284,15 +1284,15 @@ public:
 	ListwmItem* Find(ListWM &list, Token *what, Token *what2 = NULL);
 
 	// Create a new identifier by appending the word
-	Token *AppendIdentifier(Token *source, const char *word, const wchar_t *w_word, int len);
-	void AppendIdentifier(TokenStr &source, const char *word, const wchar_t *w_word, int len);
+	Token *AppendIdentifier(Token *source, const char *word, const wchar_t *w_word, size_t len);
+	void AppendIdentifier(TokenStr &source, const char *word, const wchar_t *w_word, size_t len);
 	void AppendIdentifier(TokenStr &source, TokenStr &append);
 	// Split identifier to 2 parts
 	void SplitIdentifier(Token *source, Token *first, Token *second);
 
 	// Add /* */ comments around the tokens
 	void Comment(Token *first, Token *last = NULL);
-	void Comment(const char *word, const wchar_t *w_word, int len, Token *first, Token *last);
+	void Comment(const char *word, const wchar_t *w_word, size_t len, Token *first, Token *last);
 	void CommentNoSpaces(Token *first, Token *last = NULL);
 
 	// Check whether the token is valid column/table alias
