+++ sqlparser/str.h
@@ -46,7 +46,7 @@ public:
 	static const char* SkipUntil(const char *input, char ch);
 
 	// Get copy of the input string
-	static char* GetCopy(const char *input, int size);
+	static char* GetCopy(const char *input, size_t size);
 	static char* GetCopy(const char *input);
 
 	// Replace character in string (returns the same string)
